move utility func test to utility tests
